kbuild: move "_all" target out of $(KBUILD_SRC) conditional
The first "_all" occurrence around line 120 is only visible when
KBUILD_SRC is unset.

If O=... is specified, the working directory is relocated, then the
only second occurrence around line 193 is visible, that is not set
to PHONY.

Move the first one to an always visible place.  This clarifies "_all"
is our default target and it is always set to PHONY.

 Conflicts:
	Makefile

Change-Id: Ief56843be4eb0c9faeedf34a142875f312321d3e
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Albert I <krascgq@outlook.co.id>